TST: Move some consistency rolling tests to misc rolling functions
